+++ app/Http/Controllers/CategoriaController.php
@@ -42,7 +42,12 @@ class CategoriaController extends Controller
         if(!Auth::check())
         {
             return redirect()->to('login');
-        }    
+        }  
+        if(!$this-> autorizar(Auth::user()))
+        {
+            return back();            
+        }
+       
         return view('Categoria.create');
         //
     }
@@ -56,7 +61,11 @@ class CategoriaController extends Controller
         {
             return redirect()->to('login');
         }
-    
+        if(!$this-> autorizar(Auth::user()))
+        {
+            return back();            
+        }
+     
         $validacion=$request->validate(['nombre'=>'required|max:50']);
         $this->_repository->Store((object)$request->all());
 
@@ -79,7 +88,11 @@ class CategoriaController extends Controller
         {
             return redirect()->to('login');
         }
-    
+        if(!$this-> autorizar(Auth::user()))
+        {
+            return back();            
+        }
+     
         $categoria=$this->_repository->Find($id);
         $data=[    
             'categoria'=>$categoria
@@ -96,7 +109,11 @@ class CategoriaController extends Controller
         {
             return redirect()->to('login');
         }
-    
+        if(!$this-> autorizar(Auth::user()))
+        {
+            return back();            
+        }
+     
         $validacion=$request->validate(['nombre'=>'required|max:50']);
         $this->_repository->Update($id,(object)$request->all());
         return redirect()->to(url('/categorias'));        
@@ -112,7 +129,11 @@ class CategoriaController extends Controller
         {
             return redirect()->to('login');
         }
-    
+        if(!$this-> autorizar(Auth::user()))
+        {
+            return back();            
+        }
+     
         $this->_repository->Delete($id);
         return redirect()->to(url('/categorias'));    
         //
